test(cli): Move tests doing file i/o to integration tests
